refactor(plugin_runner): split modules, introduce feature target
