+++ crates/swc/Cargo.toml
@@ -76,7 +76,7 @@ swc_ecma_visit = {version = "0.56.0", path = "../swc_ecma_visit"}
 swc_ecmascript = {version = "0.133.0", path = "../swc_ecmascript"}
 swc_error_reporters = {version = "0.1.0", path = "../swc_error_reporters"}
 swc_node_comments = {version = "0.4.0", path = "../swc_node_comments"}
-swc_plugin_runner = {version = "0.42.0", path = "../swc_plugin_runner", optional = true}
+swc_plugin_runner = {version = "0.42.0", path = "../swc_plugin_runner", optional = true, default-features = false}
 swc_visit = {version = "0.3.0", path = "../swc_visit"}
 tracing = "0.1.32"
